[refactor] 내비게이션 구조 변경
- Single Activity를 LoginActivity와 MainActivity로 분리함. 패키지 또한
animal과 group을 manage로 통합했고 추후 확장 시 feed와 navigation을 추가
할 예정. 의존성은 MainActivity에서 모듈로 구분.
