+++ mobile/app/src/main/java/com/lacuc/pets/ui/MainActivity.kt
@@ -1,9 +1,10 @@
-package com.lacuc.pets
+package com.lacuc.pets.ui
 
 import android.os.Bundle
-import dagger.android.support.DaggerAppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
+import com.lacuc.pets.R
 
-class MainActivity : DaggerAppCompatActivity() {
+class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
