+++ mobile/app/src/main/java/com/lacuc/pets/ui/login/LoginActivity.kt
@@ -1,0 +1,12 @@
+package com.lacuc.pets.ui.login
+
+import android.os.Bundle
+import com.lacuc.pets.R
+import dagger.android.support.DaggerAppCompatActivity
+
+class LoginActivity : DaggerAppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_login)
+    }
+}
