+++ mobile/app/src/main/java/com/lacuc/pets/ui/login/LoginModule.kt
@@ -1,0 +1,13 @@
+package com.lacuc.pets.ui.login
+
+import com.lacuc.pets.ui.login.signup.SignUpFragment
+import dagger.Module
+import dagger.android.ContributesAndroidInjector
+
+@Module
+abstract class LoginModule {
+
+    @ContributesAndroidInjector
+    abstract fun getSignUpFragment(): SignUpFragment
+
+}
