+++ mobile/app/src/main/java/com/lacuc/pets/ui/manage/animal/detail/memo/AddMemoFragment.kt
@@ -1,4 +1,4 @@
-package com.lacuc.pets.ui.animal.detail.medical
+package com.lacuc.pets.ui.manage.animal.detail.memo
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,20 +7,21 @@ import android.view.ViewGroup
 import androidx.fragment.app.viewModels
 import androidx.navigation.NavController
 import androidx.navigation.fragment.findNavController
-import androidx.navigation.ui.setupWithNavController
 import com.lacuc.pets.ViewModelFactory
-import com.lacuc.pets.databinding.FragmentAddMedicalBinding
+import com.lacuc.pets.databinding.FragmentAddMemoBinding
+import com.lacuc.pets.util.setupWithNavController
 import dagger.android.support.DaggerFragment
 import javax.inject.Inject
 
-class AddMedicalFragment : DaggerFragment() {
-    private var _binding: FragmentAddMedicalBinding? = null
+class AddMemoFragment : DaggerFragment() {
+
+    private var _binding: FragmentAddMemoBinding? = null
     private val binding get() = _binding!!
 
     @Inject
     lateinit var viewModelFactory: ViewModelFactory
 
-    private val viewModel: AddMedicalViewModel by viewModels { viewModelFactory }
+    private val viewModel: AddMemoViewModel by viewModels { viewModelFactory }
 
     private val navController: NavController by lazy { findNavController() }
 
@@ -29,7 +30,7 @@ class AddMedicalFragment : DaggerFragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentAddMedicalBinding.inflate(inflater, container, false).apply {
+        _binding = FragmentAddMemoBinding.inflate(inflater, container, false).apply {
             lifecycleOwner = viewLifecycleOwner
             vm = viewModel
         }
@@ -39,14 +40,14 @@ class AddMedicalFragment : DaggerFragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        binding.toolbarAddMedical.setupWithNavController(navController)
+        binding.toolbarAddMemo.setupWithNavController(navController)
 
         setOnCompleteEventObserver()
     }
 
     private fun setOnCompleteEventObserver() {
         viewModel.completeEvent.observe(viewLifecycleOwner) {
-            navController.previousBackStackEntry?.savedStateHandle?.set("onCompleteEvent", 1)
+            navController.previousBackStackEntry?.savedStateHandle?.set("onCompleteEvent", 2)
             navController.navigateUp()
         }
     }
