+++ mobile/app/src/main/java/com/lacuc/pets/ui/manage/group/choose/ChooseGroupViewModel.kt
@@ -1,4 +1,4 @@
-package com.lacuc.pets.ui.group.choose
+package com.lacuc.pets.ui.manage.group.choose
 
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
