+++ mobile/app/src/main/res/layout/fragment_add_memo.xml
@@ -2,13 +2,13 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".ui.animal.detail.memo.AddMemoFragment">
+    tools:context=".ui.manage.animal.detail.memo.AddMemoFragment">
 
     <data>
 
         <variable
             name="vm"
-            type="com.lacuc.pets.ui.animal.detail.memo.AddMemoViewModel" />
+            type="com.lacuc.pets.ui.manage.animal.detail.memo.AddMemoViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
